refactor: Trainers.js에서 modal분리
